requirements.txt
Flask==3.0.3
langchain==0.2.16
google-cloud-aiplatform==1.65.0
langchain-google-genai==1.0.10
faiss-cpu==1.8.0.post1
networkx==3.3
pandas==2.2.2
langchain-community==0.2.16
pyodbc==5.1.0

# Env management and CORS
python-dotenv==1.0.1
Flask-Cors==4.0.1

# Optional provider packages (used if selected via LLM_PROVIDER)
langchain-openai==0.1.22
langchain-anthropic==0.1.19
